Pin invoke to packages and move test functions to tests
